Uploading profile picture and profile data to Firebase

+++ TwitterClone/Models/TwitterUser.swift
@@ -12,8 +12,8 @@ struct TwitterUser: Codable {
     let id: String
     var displayName: String = ""
     var username: String = ""
-    var followersCount: Double = 0
-    var followingCount: Double = 0
+    var followersCount: Int = 0
+    var followingCount: Int = 0
     var createdOn: Date = Date()
     var bio: String = ""
     var avatarPath: String = ""
